add more feature to the generator and regenreate

+++ src/main/java/org/bshg/blogssystem/sprocess/admin/impl/DeleteAdminProcessImpl.java
@@ -12,13 +12,13 @@ import org.bshg.blogssystem.sprocess.userprofile.facade.DeleteUserProfileProcess
 import org.bshg.blogssystem.zutils.sprocess.impl.processes.AbstractDeleteProcessImpl;
 
 public class DeleteAdminProcessImpl extends AbstractDeleteProcessImpl<Admin, AdminService> implements DeleteAdminProcess {
-    public DeleteAdminProcessImpl(AdminService service, UserProfileService userProfileService, NotificationService notificationService, CommentService commentService, LikeDislikeService likeDislikeService, PostService postService) {
+    public DeleteAdminProcessImpl(AdminService service, PostService postService, LikeDislikeService likeDislikeService, NotificationService notificationService, UserProfileService userProfileService, CommentService commentService) {
         super(service);
-        this.userProfileService = userProfileService;
-        this.notificationService = notificationService;
-        this.commentService = commentService;
-        this.likeDislikeService = likeDislikeService;
         this.postService = postService;
+        this.likeDislikeService = likeDislikeService;
+        this.notificationService = notificationService;
+        this.userProfileService = userProfileService;
+        this.commentService = commentService;
     }
 
     @Override
@@ -29,13 +29,19 @@ public class DeleteAdminProcessImpl extends AbstractDeleteProcessImpl<Admin, Adm
 
     public void deleteByProfile(UserProfile profile) {
         if (profile != null && profile.getId() != null) {
+            Admin found = this.service.findByProfileId(profile.getId());
+            if (found == null) return;
+            this.deleteAssociatedList(found);
             service.deleteByProfileId(profile.getId());
         }
     }
 
     @Override
     public void deleteAssociatedList(Admin item) {
-        DeleteUserProfileProcess.deleteByProfile(item);
+        deleteLikeDislikeProcess.deleteByAdmin(item);
+        deleteNotificationProcess.deleteByAdmin(item);
+        deletePostProcess.deleteByAdmin(item);
+        deleteCommentProcess.deleteByAdmin(item);
     }
 
     @Override
@@ -43,25 +49,11 @@ public class DeleteAdminProcessImpl extends AbstractDeleteProcessImpl<Admin, Adm
         configure(Admin.class);
     }
 
-    private final UserProfileService userProfileService;
-    private DeleteUserProfileProcess deleteUserProfileProcess;
+    private final PostService postService;
+    private DeletePostProcess deletePostProcess;
 
-    public void setDeleteUserProfileProcess(DeleteUserProfileProcess value) {
-        this.deleteUserProfileProcess = value;
-    }
-
-    private final NotificationService notificationService;
-    private DeleteNotificationProcess deleteNotificationProcess;
-
-    public void setDeleteNotificationProcess(DeleteNotificationProcess value) {
-        this.deleteNotificationProcess = value;
-    }
-
-    private final CommentService commentService;
-    private DeleteCommentProcess deleteCommentProcess;
-
-    public void setDeleteCommentProcess(DeleteCommentProcess value) {
-        this.deleteCommentProcess = value;
+    public void setDeletePostProcess(DeletePostProcess value) {
+        this.deletePostProcess = value;
     }
 
     private final LikeDislikeService likeDislikeService;
@@ -71,10 +63,24 @@ public class DeleteAdminProcessImpl extends AbstractDeleteProcessImpl<Admin, Adm
         this.deleteLikeDislikeProcess = value;
     }
 
-    private final PostService postService;
-    private DeletePostProcess deletePostProcess;
+    private final NotificationService notificationService;
+    private DeleteNotificationProcess deleteNotificationProcess;
 
-    public void setDeletePostProcess(DeletePostProcess value) {
-        this.deletePostProcess = value;
+    public void setDeleteNotificationProcess(DeleteNotificationProcess value) {
+        this.deleteNotificationProcess = value;
+    }
+
+    private final UserProfileService userProfileService;
+    private DeleteUserProfileProcess deleteUserProfileProcess;
+
+    public void setDeleteUserProfileProcess(DeleteUserProfileProcess value) {
+        this.deleteUserProfileProcess = value;
+    }
+
+    private final CommentService commentService;
+    private DeleteCommentProcess deleteCommentProcess;
+
+    public void setDeleteCommentProcess(DeleteCommentProcess value) {
+        this.deleteCommentProcess = value;
     }
 }
