+++ src/main/java/org/bshg/blogssystem/sprocess/post/impl/DeletePostProcessImpl.java
@@ -14,15 +14,17 @@ import org.bshg.blogssystem.sprocess.post.facade.DeletePostProcess;
 import org.bshg.blogssystem.sprocess.tag.facade.DeleteTagProcess;
 import org.bshg.blogssystem.zutils.sprocess.impl.processes.AbstractDeleteProcessImpl;
 
+import java.util.List;
+
 public class DeletePostProcessImpl extends AbstractDeleteProcessImpl<Post, PostService> implements DeletePostProcess {
-    public DeletePostProcessImpl(PostService service, CategoryService categoryService, AdminService adminService, CommentService commentService, LikeDislikeService likeDislikeService, TagService tagService, CostumerService costumerService) {
+    public DeletePostProcessImpl(PostService service, TagService tagService, LikeDislikeService likeDislikeService, CategoryService categoryService, CostumerService costumerService, CommentService commentService, AdminService adminService) {
         super(service);
-        this.categoryService = categoryService;
-        this.adminService = adminService;
-        this.commentService = commentService;
-        this.likeDislikeService = likeDislikeService;
         this.tagService = tagService;
+        this.likeDislikeService = likeDislikeService;
+        this.categoryService = categoryService;
         this.costumerService = costumerService;
+        this.commentService = commentService;
+        this.adminService = adminService;
     }
 
     @Override
@@ -33,60 +35,41 @@ public class DeletePostProcessImpl extends AbstractDeleteProcessImpl<Post, PostS
 
     public void deleteByAdmin(Admin admin) {
         if (admin != null && admin.getId() != null) {
+            List<Post> found = this.service.findByAdminId(admin.getId());
+            if (found == null) return;
+            found.forEach(this::deleteAssociatedList);
             service.deleteByAdminId(admin.getId());
         }
     }
 
     public void deleteByCostumer(Costumer costumer) {
         if (costumer != null && costumer.getId() != null) {
+            List<Post> found = this.service.findByCostumerId(costumer.getId());
+            if (found == null) return;
+            found.forEach(this::deleteAssociatedList);
             service.deleteByCostumerId(costumer.getId());
         }
     }
 
     public void deleteByCategory(Category category) {
         if (category != null && category.getId() != null) {
+            List<Post> found = this.service.findByCategoryId(category.getId());
+            if (found == null) return;
+            found.forEach(this::deleteAssociatedList);
             service.deleteByCategoryId(category.getId());
         }
     }
 
     @Override
     public void deleteAssociatedList(Post item) {
-        DeleteAdminProcess.deleteByAdmin(item);
-        DeleteCostumerProcess.deleteByCostumer(item);
-        DeleteCategoryProcess.deleteByCategory(item);
+        deleteTagProcess.deleteByPost(item);
+        deleteCommentProcess.deleteByPost(item);
+        deleteLikeDislikeProcess.deleteByPost(item);
     }
 
     @Override
     public void configure() {
         configure(Post.class);
-    }
-
-    private final CategoryService categoryService;
-    private DeleteCategoryProcess deleteCategoryProcess;
-
-    public void setDeleteCategoryProcess(DeleteCategoryProcess value) {
-        this.deleteCategoryProcess = value;
-    }
-
-    private final AdminService adminService;
-    private DeleteAdminProcess deleteAdminProcess;
-
-    public void setDeleteAdminProcess(DeleteAdminProcess value) {
-        this.deleteAdminProcess = value;
-    }
-
-    private final CommentService commentService;
-    private DeleteCommentProcess deleteCommentProcess;
-
-    public void setDeleteCommentProcess(DeleteCommentProcess value) {
-        this.deleteCommentProcess = value;
-    }
-
-    private final LikeDislikeService likeDislikeService;
-    private DeleteLikeDislikeProcess deleteLikeDislikeProcess;
-
-    public void setDeleteLikeDislikeProcess(DeleteLikeDislikeProcess value) {
-        this.deleteLikeDislikeProcess = value;
     }
 
     private final TagService tagService;
@@ -96,10 +79,38 @@ public class DeletePostProcessImpl extends AbstractDeleteProcessImpl<Post, PostS
         this.deleteTagProcess = value;
     }
 
+    private final LikeDislikeService likeDislikeService;
+    private DeleteLikeDislikeProcess deleteLikeDislikeProcess;
+
+    public void setDeleteLikeDislikeProcess(DeleteLikeDislikeProcess value) {
+        this.deleteLikeDislikeProcess = value;
+    }
+
+    private final CategoryService categoryService;
+    private DeleteCategoryProcess deleteCategoryProcess;
+
+    public void setDeleteCategoryProcess(DeleteCategoryProcess value) {
+        this.deleteCategoryProcess = value;
+    }
+
     private final CostumerService costumerService;
     private DeleteCostumerProcess deleteCostumerProcess;
 
     public void setDeleteCostumerProcess(DeleteCostumerProcess value) {
         this.deleteCostumerProcess = value;
+    }
+
+    private final CommentService commentService;
+    private DeleteCommentProcess deleteCommentProcess;
+
+    public void setDeleteCommentProcess(DeleteCommentProcess value) {
+        this.deleteCommentProcess = value;
+    }
+
+    private final AdminService adminService;
+    private DeleteAdminProcess deleteAdminProcess;
+
+    public void setDeleteAdminProcess(DeleteAdminProcess value) {
+        this.deleteAdminProcess = value;
     }
 }
